+++ apps/native-component-list/src/screens/ExpoMaps/MapsScreen.tsx
@@ -70,6 +70,14 @@ export const MapsScreens = Platform.select({
       },
     },
     {
+      name: 'Polygon',
+      route: 'expo-maps/polygon',
+      options: {},
+      getComponent() {
+        return optionalRequire(() => require('./google/MapsPolygonScreen'));
+      },
+    },
+    {
       name: 'Circle',
       route: 'expo-maps/circle',
       options: {},
@@ -181,6 +189,14 @@ export const MapsScreens = Platform.select({
       options: {},
       getComponent() {
         return optionalRequire(() => require('./apple/MapsPolylineScreen'));
+      },
+    },
+    {
+      name: 'Polygon',
+      route: 'expo-maps/polygon',
+      options: {},
+      getComponent() {
+        return optionalRequire(() => require('./apple/MapsPolygonScreen'));
       },
     },
     {